新增文件下载功能

+++ src/main/java/com/dyrnq/bdcm/model/Artifact.java
@@ -39,6 +39,6 @@ public class Artifact {
     private Integer autoJob;
 
     @Column("current_job_id")
-    private Integer currentJobId;
+    private Long currentJobId;
 
 }
